feat(remove): added confirmation prompt

+++ src/modules/remove.cpp
@@ -12,6 +12,13 @@ int remove_config() {
     return 1;
   }
 
+  if (!ctx->forced) {
+    std::string choice;
+    std::cout << "Do you really want to delete the config? (y/N) ";
+    std::getline(std::cin, choice);
+    if (choice != "y" && choice != "Y") return 0;
+  }
+
   std::error_code err;
   if (!std::filesystem::remove_all(path, err)) {
     std::cout << "Oopsies, a error happened :( \n" << err.message() << std::endl;
